+++ modules/040-control-plane-manager/images/control-plane-manager/src/internal/controllers/control-plane-node/controller_test.go
@@ -38,8 +38,8 @@ import (
 
 	controlplanev1alpha1 "control-plane-manager/api/v1alpha1"
 	"control-plane-manager/internal/constants"
-	"control-plane-manager/internal/cpnplanner"
-	"control-plane-manager/internal/cpnreconcile"
+	"control-plane-manager/internal/cpn/cpnplanner"
+	"control-plane-manager/internal/cpn/cpnreconcile"
 )
 
 var (
