+++ modules/040-control-plane-manager/images/control-plane-manager/src/internal/controllers/virtual-control-plane-node/controller.go
@@ -30,8 +30,8 @@ import (
 
 	controlplanev1alpha1 "control-plane-manager/api/v1alpha1"
 	"control-plane-manager/internal/constants"
-	"control-plane-manager/internal/cpnplanner"
-	"control-plane-manager/internal/cpnreconcile"
+	"control-plane-manager/internal/cpn/cpnplanner"
+	"control-plane-manager/internal/cpn/cpnreconcile"
 )
 
 func BuildController(mgr manager.Manager) error {
