+++ modules/040-control-plane-manager/images/control-plane-manager/src/internal/cpn/cpnplanner/plan.go
