+++ modules/040-control-plane-manager/images/control-plane-manager/src/internal/cpn/cpnplanner/steps_common.go
