+++ modules/040-control-plane-manager/images/control-plane-manager/src/internal/cpn/cpnreconcile/reconciler.go
@@ -30,7 +30,7 @@ import (
 
 	controlplanev1alpha1 "control-plane-manager/api/v1alpha1"
 	"control-plane-manager/internal/constants"
-	"control-plane-manager/internal/cpnplanner"
+	"control-plane-manager/internal/cpn/cpnplanner"
 	"control-plane-manager/internal/operations"
 )
